+++ src/main/java/io/vinyldns/java/model/zone/ZoneResponse.java
@@ -13,33 +13,130 @@
  */
 package io.vinyldns.java.model.zone;
 
+import org.joda.time.DateTime;
+
+import java.util.Objects;
+
 public class ZoneResponse {
-  private final Zone zone;
+    private Zone zone;
+    private String userId;
+    private ZoneChangeType changeType;
+    private ZoneChangeStatus status = ZoneChangeStatus.Pending;
+    private DateTime created;
+    private String systemMessage; //optional
+    private String id;
 
-  public ZoneResponse(Zone zone) {
-    this.zone = zone;
-  }
+    public ZoneResponse() {}
 
-  public Zone getZone() {
-    return zone;
-  }
+    public ZoneResponse(Zone zone,
+                        String userId,
+                        ZoneChangeType changeType,
+                        ZoneChangeStatus status,
+                        DateTime created,
+                        String systemMessage,
+                        String id) {
+        this.zone = zone;
+        this.userId = userId;
+        this.changeType = changeType;
+        this.status = status;
+        this.created = created;
+        this.systemMessage = systemMessage;
+        this.id = id;
+    }
 
-  @Override
-  public String toString() {
-    return zone.toString();
-  }
+    public Zone getZone() {
+        return zone;
+    }
 
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
+    public void setZone(Zone zone) {
+        this.zone = zone;
+    }
 
-    ZoneResponse that = (ZoneResponse) o;
-    return zone.equals(that.zone);
-  }
+    public String getUserId() {
+        return userId;
+    }
 
-  @Override
-  public int hashCode() {
-    return zone.hashCode();
-  }
+    public void setUserId(String userId) {
+        this.userId = userId;
+    }
+
+    public ZoneChangeType getChangeType() {
+        return changeType;
+    }
+
+    public void setChangeType(ZoneChangeType changeType) {
+        this.changeType = changeType;
+    }
+
+    public ZoneChangeStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(ZoneChangeStatus status) {
+        this.status = status;
+    }
+
+    public DateTime getCreated() {
+        return created;
+    }
+
+    public void setCreated(DateTime created) {
+        this.created = created;
+    }
+
+    public String getSystemMessage() {
+        return systemMessage;
+    }
+
+    public void setSystemMessage(String systemMessage) {
+        this.systemMessage = systemMessage;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    @Override
+    public String toString() {
+        return "ZoneResponse{" +
+                "zone=" + zone +
+                ", userId='" + userId + '\'' +
+                ", changeType=" + changeType +
+                ", status=" + status +
+                ", created=" + created +
+                ", systemMessage='" + systemMessage + '\'' +
+                ", id='" + id + '\'' +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof ZoneResponse)) return false;
+        ZoneResponse that = (ZoneResponse) o;
+        return Objects.equals(getZone(), that.getZone()) &&
+                Objects.equals(getUserId(), that.getUserId()) &&
+                getChangeType() == that.getChangeType() &&
+                getStatus() == that.getStatus() &&
+                Objects.equals(getCreated(), that.getCreated()) &&
+                Objects.equals(getSystemMessage(), that.getSystemMessage()) &&
+                Objects.equals(getId(), that.getId());
+    }
+
+    @Override
+    public int hashCode() {
+        int result = zone.hashCode();
+        result = 31 * result + (userId != null ? userId.hashCode() : 0);
+        result = 31 * result + (changeType != null ? changeType.hashCode() : 0);
+        result = 31 * result + (status != null ? status.hashCode() : 0);
+        result = 31 * result + (created != null ? created.hashCode() : 0);
+        result = 31 * result + (systemMessage != null ? systemMessage.hashCode() : 0);
+        result = 31 * result + (id != null ? id.hashCode() : 0);
+        return result;
+    }
+
 }
