+++ src/test/java/io/vinyldns/java/VinylDNSClientTest.java
@@ -122,8 +122,44 @@ public class VinylDNSClientTest {
   }
 
   @Test
+  public void createZoneSucess() {
+    Zone newZone = new Zone(
+                    "newid",
+                    "name",
+                    "some@company.com",
+                    ZoneStatus.Active,
+                    new DateTime(),
+                    new DateTime(),
+                    testZoneConnection1,
+                    null,
+                    true,
+                    testZoneAcl1,
+                    "group",
+                    null);
+
+    ZoneResponse response = new ZoneResponse(
+            newZone,"someUserId", ZoneChangeType.Create,ZoneChangeStatus.Pending, new DateTime(), "", "1234");
+
+    wireMockServer.stubFor(
+        post(urlEqualTo("/zones"))
+            .willReturn(
+            aResponse()
+                    .withStatus(202)
+                    .withBody(newZone)
+                    .withHeader("Content-Type", "application/json")
+                    .withBody(response)));
+
+    VinylDNSResponse<ZoneResponse> vinylDNSResponse =
+            client.createZone(newZone);
+
+    assertTrue(vinylDNSResponse instanceof ResponseMarker.Success);
+    assertEquals(vinylDNSResponse.getStatusCode(), 202);
+    assertEquals(vinylDNSResponse.getValue().getZone(), newZone);
+  }
+
+  @Test
   public void getZoneSuccess() {
-    String response = client.gson.toJson(new ZoneResponse(testZone1));
+    String response = client.gson.toJson(new GetZoneResponse(testZone1));
 
     wireMockServer.stubFor(
         get(urlEqualTo("/zones/" + testZone1.getId()))
@@ -133,7 +169,7 @@ public class VinylDNSClientTest {
                     .withHeader("Content-Type", "application/json")
                     .withBody(response)));
 
-    VinylDNSResponse<ZoneResponse> vinylDNSResponse =
+    VinylDNSResponse<GetZoneResponse> vinylDNSResponse =
         client.getZone(new ZoneRequest(testZone1.getId()));
 
     assertTrue(vinylDNSResponse instanceof ResponseMarker.Success);
@@ -147,12 +183,120 @@ public class VinylDNSClientTest {
         get(urlEqualTo("/zones/" + zoneId))
             .willReturn(aResponse().withStatus(404).withBody("not found")));
 
-    VinylDNSResponse<ZoneResponse> vinylDNSResponse = client.getZone(new ZoneRequest(zoneId));
+    VinylDNSResponse<GetZoneResponse> vinylDNSResponse = client.getZone(new ZoneRequest(zoneId));
 
     assertTrue(vinylDNSResponse instanceof ResponseMarker.Failure);
     assertEquals(vinylDNSResponse.getStatusCode(), 404);
     assertEquals(vinylDNSResponse.getMessageBody(), "not found");
     assertNull(vinylDNSResponse.getValue());
+  }
+
+  @Test
+  public void updateZoneSucess() {
+    Zone newZone = new Zone(
+            "newid",
+            "name123",
+            "some@company.com",
+            ZoneStatus.Active,
+            new DateTime(),
+            new DateTime(),
+            testZoneConnection1,
+            null,
+            true,
+            testZoneAcl1,
+            "group",
+            null);
+
+    ZoneResponse response = new ZoneResponse(
+            newZone,"someUserId", ZoneChangeType.Update,ZoneChangeStatus.Pending, new DateTime(), "", "1234");
+
+    wireMockServer.stubFor(
+            put(urlEqualTo("/zones/" + newZone.getId()))
+                    .willReturn(
+                            aResponse()
+                                    .withStatus(202)
+                                    .withBody(newZone)
+                                    .withHeader("Content-Type", "application/json")
+                                    .withBody(response)));
+
+    VinylDNSResponse<ZoneResponse> vinylDNSResponse =
+            client.updateZone(newZone);
+
+    assertTrue(vinylDNSResponse instanceof ResponseMarker.Success);
+    assertEquals(vinylDNSResponse.getStatusCode(), 202);
+    assertEquals(vinylDNSResponse.getValue().getZone(), newZone);
+  }
+
+  @Test
+  public void deleteZoneSuccess() {
+    ZoneResponse response = new ZoneResponse(
+            testZone1,"someUserId", ZoneChangeType.Delete,ZoneChangeStatus.Pending, new DateTime(), "", "1234");
+
+    wireMockServer.stubFor(
+            delete(urlEqualTo("/zones/" + testZone1.getId()))
+                    .willReturn(
+                            aResponse()
+                                    .withStatus(202)
+                                    .withHeader("Content-Type", "application/json")
+                                    .withBody(response)));
+
+    VinylDNSResponse<ZoneResponse> vinylDNSResponse =
+            client.deleteZone(new ZoneRequest(testZone1.getId()));
+
+    assertTrue(vinylDNSResponse instanceof ResponseMarker.Success);
+    assertEquals(vinylDNSResponse.getStatusCode(), 202);
+    assertEquals(vinylDNSResponse.getValue().getZone(), testZone1);
+  }
+
+  @Test
+  public void listZoneChangesSuccess() {
+
+    String recordNameFilter = "someFilter";
+    String startFrom = "someStart";
+    int maxItems = 55;
+    List<ZoneResponse> zoneChanges = Collections.singletonList(
+            testZone1,"someUserId", ZoneChangeType.Update,ZoneChangeStatus.Pending, new DateTime(), "", "1234");
+
+    ListZoneChangesResponse listZoneChangesResponse = new ListZoneChangesResponse(zoneChanges,startFrom, "nextId", maxItems, recordNameFilter)
+
+    wireMockServer.stubFor(
+            get(urlEqualTo("/zones/" + testZone1.getId() + "/changes?(.*)"))
+                    .withQueryParam("recordNameFilter", equalTo(recordNameFilter))
+                    .withQueryParam("startFrom", equalTo(startFrom))
+                    .withQueryParam("maxItems", equalTo(String.valueOf(maxItems)))
+                    .willReturn(
+                            aResponse()
+                                    .withStatus(200)
+                                    .withHeader("Content-Type", "application/json")
+                                    .withBody(listZoneChangesResponse)));
+
+    VinylDNSResponse<ListZoneChangesResponse> vinylDNSResponse =
+            client.listZoneChanges(new ListZoneChangesRequest(testZone1.getId(), startFrom, maxItems));
+
+
+    assertTrue(vinylDNSResponse instanceof ResponseMarker.Success);
+    assertEquals(vinylDNSResponse.getStatusCode(), 200);
+    assertEquals(vinylDNSResponse.getValue().getZoneChanges(), zoneChanges);
+  }
+
+  @Test
+  public void syncZoneSuccess() {
+    String response = client.gson.toJson(new GetZoneResponse(testZone1));
+
+    wireMockServer.stubFor(
+            post(urlEqualTo("/zones/" + testZone1.getId() + "/sync"))
+                    .willReturn(
+                            aResponse()
+                                    .withStatus(202)
+                                    .withHeader("Content-Type", "application/json")
+                                    .withBody(response)));
+
+    VinylDNSResponse<ZoneResponse> vinylDNSResponse =
+            client.deleteZone(new ZoneRequest(testZone1.getId()));
+
+    assertTrue(vinylDNSResponse instanceof ResponseMarker.Success);
+    assertEquals(vinylDNSResponse.getStatusCode(), 200);
+    assertEquals(vinylDNSResponse.getValue().getZone(), testZone1);
   }
 
   @Test
